rename one-function file as per suggestion
